refactor(sandbox-git): extract refspec + push-guard pure core
Move pushDestinationBranch, isDeleteRefspec and the git_push force/delete/
default-branch decision into sandbox-git/core/refspec.ts as evaluatePushGuard.
100% branch coverage; git_push delegates to the extracted guard. The frozen
68KB call-shape suite passes unmodified.
